DocExt/requirements.txt
streamlit
ultralytics
opencv-python-headless
Pillow
PyMuPDF
easyocr
numpy<2.0